requirements.txt
flask==3.0.3
psycopg2-binary==2.9.9
werkzeug==3.0.3
apscheduler==3.10.4
groq==0.9.0
google-generativeai==0.7.2
requests==2.32.3
beautifulsoup4==4.12.3
PyMuPDF==1.24.5
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.2
pytesseract==0.3.13
Pillow==10.4.0
lxml==5.2.2
bcrypt==4.1.3
boto3==1.34.69
playwright==1.44.0
